backend/requirements.txt
fastapi
uvicorn
torch
transformers
pillow
python-multipart
pydantic
